# backend/swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http"
  ],
  "swagger": "2.0",
  "info": {
    "description": "BMSTU Web Course 2020 project",
    "title": "Forums API",
    "version": "1.0.0"
  },
  "host": "localhost:5000",
  "paths": {
...
    "/api/v1/post/:id/details": {
      "get": {
        "description": "Получение информации о ветке обсуждения по его имени.",
        "tags": [
          "posts"
        ],
        "summary": "Получение информации о ветке обсуждения",
        "operationId": "postGetOne",
        "parameters": [
          {
            "type": "number",
            "format": "int64",
            "description": "Идентификатор сообщения.",
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "items": {
              "enum": [
                "user",
                "forum",
                "thread"
              ]
            },
            "description": "Включение полной информации о соответвующем объекте сообщения. Если тип объекта не указан, то полная информация об этих объектах не передаётся.",
            "name": "related",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Информация о ветке обсуждения.",
            "schema": {
              "$ref": "#/definitions/PostFull"
            }
          },
          "404": {
            "description": "Ветка обсуждения отсутсвует в форуме.",
            "schema": {
              "$ref": "#/responses/Error"
            }
          }
        }
      },
      "post": {
        "description": "Изменение сообщения на форуме. Если сообщение поменяло текст, то оно должно получить отметку `isEdited`.",
        "tags": [
          "posts"
        ],
        "summary": "Изменение сообщения",
        "operationId": "postUpdate",
        "parameters": [
          {
            "type": "number",
            "format": "int64",
            "description": "Идентификатор сообщения.",
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "description": "Изменения сообщения.",
            "name": "post",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/PostUpdate"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Информация о сообщении.",
            "schema": {
              "$ref": "#/definitions/Post"
            }
          },
          "404": {
            "description": "Сообщение отсутсвует в форуме.",
            "schema": {
              "$ref": "#/responses/Error"
            }
          }
        }
      }
    },
...
    "/api/v1/thread/:slug/posts": {
      "get": {
        "description": "Получение списка сообщений в данной ветке форуме. Сообщения выводятся отсортированные по дате создания.",
        "tags": [
          "threads"
        ],
        "summary": "Сообщения данной ветви обсуждения",
        "operationId": "threadGetPosts",
        "parameters": [
          {
            "type": "string",
            "format": "identity",
            "description": "Идентификатор ветки обсуждения.",
            "name": "slug_or_id",
            "in": "path",
            "required": true
          },
          {
            "maximum": 10000,
            "minimum": 1,
            "type": "number",
            "format": "int32",
            "default": 100,
            "description": "Максимальное кол-во возвращаемых записей.",
            "name": "limit",
            "in": "query"
          },
          {
            "type": "number",
            "format": "int64",
            "description": "Идентификатор поста, после которого будут выводиться записи (пост с данным идентификатором в результат не попадает).",
            "name": "since",
            "in": "query"
          },
          {
            "enum": [
              "flat",
              "tree",
              "parent_tree"
            ],
            "type": "string",
            "default": "flat",
            "description": "Вид сортировки:\n* flat - по дате, комментарии выводятся простым списком в порядке создания;\n* tree - древовидный, комментарии выводятся отсортированные в дереве по N штук; \n* parent_tree - древовидные с пагинацией по родительским (parent_tree), на странице N родительских комментов и все комментарии прикрепленные к ним, в древвидном отображение.",
            "name": "sort",
            "in": "query"
          },
          {
            "type": "boolean",
            "description": "Флаг сортировки по убыванию.",
            "name": "desc",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Информация о сообщениях форума.",
            "schema": {
              "$ref": "#/definitions/Posts"
            }
          },
          "404": {
            "description": "Ветка обсуждения отсутсвует в форуме.",
            "schema": {
              "$ref": "#/responses/Error"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "Forum": {
      "description": "Информация о форуме",
      "type": "object",
      "required": [
        "slug",
        "title"
      ],
      "properties": {
        "posts": {
          "description": "Общее количество сообщений в форуме",
          "type": "integer",
          "format": "int64",
          "x-go-name": "Posts",
          "readOnly": true,
          "example": 200000
        },
        "slug": {
          "description": "Человекопонятный URL, уникальное поле",
          "type": "string",
          "uniqueItems": true,
          "x-go-name": "Slug",
          "example": "pirate-stories"
        },
        "threads": {
          "description": "Общее количество ветвей обсуждения в форуме",
          "type": "integer",
          "format": "int64",
          "x-go-name": "Threads",
          "readOnly": true,
          "example": 200
        },
        "title": {
          "description": "Название форума",
          "type": "string",
          "x-go-name": "Title",
          "example": "Pirate stories"
        },
        "user": {
          "description": "Nickname пользователя, создавшего форум",
          "type": "string",
          "x-go-name": "User",
          "example": "j.sparrow"
        }
      },
      "x-go-name": "ForumOutput",
      "x-go-package": "dbProject/forum/delivery/http"
    },
...
    "Post": {
      "description": "Информация о сообщении",
      "type": "object",
      "properties": {
        "author": {
          "description": "Автор сообщения\n\nformat: identity",
          "type": "string",
          "x-go-name": "Author",
          "example": "j.sparrow"
        },
        "created": {
          "description": "Дата создания",
          "type": "string",
          "format": "date-time",
          "x-go-name": "Created",
          "readOnly": true
        },
        "forum": {
          "description": "Идентификатор форума",
          "type": "string",
          "x-go-name": "ForumName",
          "readOnly": true
        },
        "id": {
          "description": "Идентификатор сообщения",
          "type": "integer",
          "format": "int64",
          "x-go-name": "Pk",
          "readOnly": true
        },
        "isEdited": {
          "description": "Истина, если сообщение было изменено",
          "type": "boolean",
          "x-go-name": "IsEdited",
          "readOnly": true
        },
        "message": {
          "description": "Текст сообщения",
          "type": "string",
          "x-go-name": "Message",
          "example": "We should be afraid of the Kraken"
        },
        "parent": {
          "description": "Идентификатор родительского сообщения",
          "type": "integer",
          "format": "int64",
          "x-go-name": "Parent"
        },
        "thread": {
          "description": "Идентификатор ветви обсуждения данного сообщения",
          "type": "integer",
          "format": "int64",
          "x-go-name": "Tid",
          "readOnly": true
        }
      },
      "x-go-package": "dbProject/forum/delivery/http"
    },
    "PostFull": {
      "description": "Полная информация о сообщени, включая связанные объекты",
      "type": "object",
      "properties": {
        "author": {
          "$ref": "#/definitions/UserOutput"
        },
        "forum": {
          "$ref": "#/definitions/Forum"
        },
        "post": {
          "$ref": "#/definitions/Post"
        },
        "thread": {
          "$ref": "#/definitions/Thread"
        }
      },
      "x-go-name": "DetailedInfo",
      "x-go-package": "dbProject/forum/delivery/http"
    },
    "PostUpdate": {
      "description": "Сообщение для обновления существующего",
      "type": "object",
      "properties": {
        "message": {
          "description": "Собственно сообщение",
          "type": "string",
          "x-go-name": "Message",
          "example": "We should be afraid of the Kraken"
        }
      },
      "x-go-name": "postInput",
      "x-go-package": "dbProject/forum/delivery/http"
    },
    "Posts": {
      "type": "object",
      "properties": {
        "posts": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Post"
          },
          "x-go-name": "Posts"
        }
      },
      "x-go-package": "dbProject/forum/delivery/http"
    },
...
    "Thread": {
      "description": "Ветка обсуждений на форуме",
      "type": "object",
      "properties": {
        "author": {
          "description": "Пользователь, создавший ветку\n\nformat: identity",
          "type": "string",
          "x-go-name": "Author",
          "example": "j.sparrow"
        },
        "created": {
          "description": "Дата создания ветки",
          "type": "string",
          "format": "date-time",
          "x-go-name": "Created",
          "example": "2017-01-01T00:00:00.000Z"
        },
        "forum": {
          "description": "Форум ветки\n\nformat: identity",
          "type": "string",
          "x-go-name": "ForumName",
          "readOnly": true,
          "example": "pirate-stories"
        },
        "id": {
          "description": "Идентификатор ветки обсуждения",
          "type": "integer",
          "format": "int64",
          "x-go-name": "Id",
          "readOnly": true,
          "example": 42
        },
        "message": {
          "description": "Описание ветки обсуждения",
          "type": "string",
          "x-go-name": "Message",
          "example": "An urgent need to reveal the hiding place of Davy Jones. Who is willing to help in this matter?"
        },
        "posts": {
          "description": "Количество сообщений в ветке",
          "type": "integer",
          "format": "int64",
          "x-go-name": "PostsCount",
          "readOnly": true,
          "example": 100
        },
        "slug": {
          "description": "Человекопонятный URL\n\nformat: identity",
          "type": "string",
          "x-go-name": "Slug",
          "readOnly": true,
          "example": "jones-cache"
        },
        "title": {
          "description": "Заголовок ветки",
          "type": "string",
          "x-go-name": "Title",
          "example": "Davy Jones cache"
        },
        "votes": {
          "description": "Кол-во голосов",
          "type": "integer",
          "format": "int64",
          "x-go-name": "Votes",
          "readOnly": true
        }
      },
      "x-go-package": "dbProject/forum/delivery/http"
    },
...
    "UserOutput": {
      "type": "object",
      "properties": {
        "about": {
          "type": "string",
          "x-go-name": "About"
        },
        "email": {
          "type": "string",
          "x-go-name": "Email"
        },
        "fullname": {
          "type": "string",
          "x-go-name": "Fullname"
        },
        "nickname": {
          "type": "string",
          "x-go-name": "Nickname"
        }
      },
      "x-go-package": "dbProject/user/delivery/http"
    },
...
  },
  "responses": {
    "Error": {
      "description": "Error response payload",
      "schema": {
        "type": "object",
        "properties": {
          "message": {
            "description": "Текстовое описание ошибки.",
            "type": "string",
            "x-go-name": "Message"
          }
        }
      }
    }
  },
  "securityDefinitions": {
    "api_key": {
      "type": "apiKey",
      "name": "KEY",
      "in": "header"
    }
  },
  "security": [
    {
      "api_key": []
    }
  ]
}